Stopped waiting for drawer to close to switch channels
