+++ app/components/channel_drawer/channel_drawer.js
@@ -7,6 +7,7 @@ import {
     BackHandler,
     InteractionManager,
     Keyboard,
+    StyleSheet,
     View
 } from 'react-native';
 
@@ -35,29 +36,20 @@ export default class ChannelDrawer extends PureComponent {
         }).isRequired,
         blurPostTextBox: PropTypes.func.isRequired,
         children: PropTypes.node,
-        channels: PropTypes.object,
-        currentChannel: PropTypes.object,
+        currentChannelId: PropTypes.string.isRequired,
         currentDisplayName: PropTypes.string,
-        channelMembers: PropTypes.object,
-        currentTeam: PropTypes.object,
+        currentTeamId: PropTypes.string.isRequired,
         currentUserId: PropTypes.string.isRequired,
         intl: PropTypes.object.isRequired,
-        myTeamMembers: PropTypes.object.isRequired,
         navigator: PropTypes.object,
+        teamsCount: PropTypes.number.isRequired,
         theme: PropTypes.object.isRequired
-    };
-
-    static defaultProps = {
-        currentTeam: {},
-        currentChannel: {}
     };
 
     state = {
         openDrawer: false,
         openDrawerOffset: DRAWER_INITIAL_OFFSET
     };
-
-    swiperIndex = 1;
 
     componentWillMount() {
         this.props.actions.getTeams();
@@ -88,40 +80,23 @@ export default class ChannelDrawer extends PureComponent {
         this.setState({openDrawer: false});
     };
 
-    handleDrawerClose = () => {
-        this.resetDrawer();
-
-        if (this.closeLeftHandle) {
-            InteractionManager.clearInteractionHandle(this.closeLeftHandle);
-            this.closeLeftHandle = null;
-        }
-
-        if (this.state.openDrawer) {
-            this.setState({openDrawer: false});
-        }
-    };
-
     handleDrawerCloseStart = () => {
-        if (!this.closeLeftHandle) {
-            this.closeLeftHandle = InteractionManager.createInteractionHandle();
+        if (this.state.openDrawer) {
+            // The state doesn't get updated if you swipe to close
+            this.setState({
+                openDrawer: false
+            });
         }
-    };
-
-    handleDrawerOpen = () => {
-        this.setState({openDrawer: true});
-        if (this.state.openDrawerOffset === DRAWER_INITIAL_OFFSET) {
-            Keyboard.dismiss();
-        }
-
-        if (this.openLeftHandle) {
-            InteractionManager.clearInteractionHandle(this.openLeftHandle);
-            this.openLeftHandle = null;
-        }
-    };
+    }
 
     handleDrawerOpenStart = () => {
-        if (!this.openLeftHandle) {
-            this.openLeftHandle = InteractionManager.createInteractionHandle();
+        Keyboard.dismiss();
+
+        if (!this.state.openDrawer) {
+            // The state doesn't get updated if you swipe to open
+            this.setState({
+                openDrawer: true
+            });
         }
     };
 
@@ -145,13 +120,16 @@ export default class ChannelDrawer extends PureComponent {
 
     openChannelDrawer = () => {
         this.props.blurPostTextBox();
-        this.setState({openDrawer: true});
+
+        this.setState({
+            openDrawer: true
+        });
     };
 
     selectChannel = (channel) => {
         const {
             actions,
-            currentChannel
+            currentChannelId
         } = this.props;
 
         const {
@@ -162,22 +140,24 @@ export default class ChannelDrawer extends PureComponent {
             viewChannel
         } = actions;
 
-        markChannelAsRead(channel.id, currentChannel.id);
-        setChannelLoading();
-        viewChannel(currentChannel.id);
         setChannelDisplayName(channel.display_name);
+        handleSelectChannel(channel.id);
+
         this.closeChannelDrawer();
+
         InteractionManager.runAfterInteractions(() => {
-            handleSelectChannel(channel.id);
+            markChannelAsRead(channel.id, currentChannelId);
+            setChannelLoading();
+            viewChannel(currentChannelId);
         });
     };
 
     joinChannel = async (channel) => {
         const {
             actions,
-            currentChannel,
+            currentChannelId,
             currentDisplayName,
-            currentTeam,
+            currentTeamId,
             currentUserId,
             intl
         } = this.props;
@@ -192,46 +172,42 @@ export default class ChannelDrawer extends PureComponent {
             viewChannel
         } = actions;
 
-        markChannelAsRead(currentChannel.id);
+        markChannelAsRead(currentChannelId);
         setChannelLoading();
-        viewChannel(currentChannel.id);
+        viewChannel(currentChannelId);
         setChannelDisplayName(channel.display_name);
 
         const displayValue = {displayName: channel.display_name};
 
+        let result;
         if (channel.type === General.DM_CHANNEL) {
-            const result = await makeDirectChannel(channel.id);
+            result = await makeDirectChannel(channel.id);
+
             if (result.error) {
                 const dmFailedMessage = {
                     id: 'mobile.open_dm.error',
                     defaultMessage: "We couldn't open a direct message with {displayName}. Please check your connection and try again."
                 };
-                setChannelDisplayName(currentDisplayName);
                 alertErrorWithFallback(intl, result.error, dmFailedMessage, displayValue);
-            } else {
-                this.closeChannelDrawer();
             }
         } else {
-            const result = await joinChannel(currentUserId, currentTeam.id, channel.id);
+            result = await joinChannel(currentUserId, currentTeamId, channel.id);
 
             if (result.error) {
                 const joinFailedMessage = {
                     id: 'mobile.join_channel.error',
                     defaultMessage: "We couldn't join the channel {displayName}. Please check your connection and try again."
                 };
-                setChannelDisplayName(currentDisplayName);
                 alertErrorWithFallback(intl, result.error, joinFailedMessage, displayValue);
-            } else {
-                this.closeChannelDrawer();
-                InteractionManager.runAfterInteractions(() => {
-                    handleSelectChannel(channel.id);
-                });
             }
         }
-    };
 
-    onPageSelected = (index) => {
-        this.swiperIndex = index;
+        if (result.error) {
+            setChannelDisplayName(currentDisplayName);
+        } else {
+            handleSelectChannel(channel.id);
+            this.closeChannelDrawer();
+        }
     };
 
     onSearchEnds = () => {
@@ -246,38 +222,30 @@ export default class ChannelDrawer extends PureComponent {
     };
 
     showTeams = () => {
-        const teamsCount = Object.keys(this.props.myTeamMembers).length;
-        if (this.swiperIndex === 1 && teamsCount > 1) {
+        if (this.props.teamsCount > 1) {
             this.refs.swiper.showTeamsPage();
-        }
-    };
-
-    resetDrawer = () => {
-        if (this.swiperIndex !== 1) {
-            this.refs.swiper.resetPage();
         }
     };
 
     renderContent = () => {
         const {
-            currentChannel,
-            currentTeam,
-            channels,
-            channelMembers,
             navigator,
-            myTeamMembers,
+            teamsCount,
             theme
         } = this.props;
-        const {openDrawerOffset} = this.state;
-        const showTeams = openDrawerOffset === DRAWER_INITIAL_OFFSET && Object.keys(myTeamMembers).length > 1;
+
+        const {
+            openDrawerOffset
+        } = this.state;
+
+        const showTeams = openDrawerOffset === DRAWER_INITIAL_OFFSET && teamsCount > 1;
 
         let teams;
         if (showTeams) {
             teams = (
-                <View style={{flex: 1, marginBottom: 10}}>
+                <View style={style.swiperContent}>
                     <TeamsList
                         closeChannelDrawer={this.closeChannelDrawer}
-                        myTeamMembers={myTeamMembers}
                         navigator={navigator}
                     />
                 </View>
@@ -285,17 +253,11 @@ export default class ChannelDrawer extends PureComponent {
         }
 
         const channelsList = (
-            <View style={{flex: 1, marginBottom: 10}}>
+            <View style={style.swiperContent}>
                 <ChannelsList
-                    currentTeam={currentTeam}
-                    currentChannel={currentChannel}
-                    channels={channels}
-                    channelMembers={channelMembers}
-                    myTeamMembers={myTeamMembers}
-                    theme={theme}
+                    navigator={navigator}
                     onSelectChannel={this.selectChannel}
                     onJoinChannel={this.joinChannel}
-                    navigator={navigator}
                     onShowTeams={this.showTeams}
                     onSearchStart={this.onSearchStart}
                     onSearchEnds={this.onSearchEnds}
@@ -305,8 +267,6 @@ export default class ChannelDrawer extends PureComponent {
 
         return (
             <Swiper
-                ref='swiper'
-                onPageSelected={this.onPageSelected}
                 openDrawerOffset={openDrawerOffset}
                 showTeams={showTeams}
                 theme={theme}
@@ -315,7 +275,7 @@ export default class ChannelDrawer extends PureComponent {
                 {channelsList}
             </Swiper>
         );
-    };
+    }
 
     render() {
         const {children} = this.props;
@@ -363,3 +323,10 @@ export default class ChannelDrawer extends PureComponent {
         );
     }
 }
+
+const style = StyleSheet.create({
+    swiperContent: {
+        flex: 1,
+        marginBottom: 10
+    }
+});
